+++ site/js/dev/ponymail_zzz.js
@@ -18,6 +18,8 @@
 
 // dealWithKeyboard: Handles what happens when you hit the escape key
 function dealWithKeyboard(e) {
+    
+    // escape key: hide composer/settings/thread
     if (e.keyCode == 27) {
         if (document.getElementById('splash').style.display == 'block') {
             document.getElementById('splash').style.display = "none"
@@ -53,6 +55,41 @@ function dealWithKeyboard(e) {
             }
         }
     }
+    // H key: show help
+    else if (e.keyCode == 72) {
+        popup("Keyboard shortcuts",
+              "<pre>\
+              <b>H:</b>Show this help menu<br/>\
+              <b>C:</b>Compose a new email to the current list<br/>\
+              <b>R:</b>Reply to the last opened email<br/>\
+              <b>S:</b>Go to the quick search bar<br/>\
+              <b>Esc:</b>Hide/collapse current email or thread<br/>\
+              </pre>\
+              You can also, in some cases, use the mouse wheel to scroll up/down the list view",
+              10
+              )
+    }
+    
+    // The key shortcuts below depend on the splash window being hidden
+    
+    // C key: compose
+    else if (e.keyCode == 67 && document.getElementById('splash').style.display != 'block') {
+        compose(null, xlist, 'new')
+    }
+    // R key: reply
+    else if (e.keyCode == 82 && document.getElementById('splash').style.display != 'block') {
+        if (openEmail() && last_opened_email) {
+            compose(last_opened_email, null, 'reply')
+        }
+    }
+    // S key: quick search
+    else if (e.keyCode == 83 && document.getElementById('splash').style.display != 'block') {
+        if (document.getElementById('q')) {
+            document.getElementById('q').focus()
+        }
+        
+    }
+    
 }