+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Hitting the escape key will now properly close threads in treeview mode
 - Fixed an issue where threads were sometimes sorted the wrong way, time-wise
 - UI shortcut improvements for high resolution users
+- Added additional keyboard shortcuts and a help menu (H key to open)
 
 ## CHANGES in 0.7b: